Update nosy list also when an issue us *updated*
Previously, only when a new issue was created the nosy list was
updated. This commit also update the nosy list when the topic is
changed on an already created issue is updated.

Note that the the nosy list is only extended: no members in the nosy
list is removed even if the topic is later removed.